build balance page and record page

+++ app/src/main/java/com/dicky/pokulator/User.java
@@ -34,7 +34,7 @@ public class User {
         return balance;
     }
 
-    public void SetBalance (int change) {
+    public void SetBalance (double change) {
         balance += change;
     }
 }
